[STABLE] Sync button tests
